add views

+++ src/main/resources/db/changelog/changes/001-init-schema.sql
@@ -1,7 +1,7 @@
 CREATE TYPE hotel.room_type AS ENUM ('Стандарт', 'Семейный', 'Люкс');
 
 CREATE TABLE IF NOT EXISTS hotel.room_types (
-    type_id SERIAL PRIMARY KEY,
+    type_id BIGSERIAL PRIMARY KEY,
     type_name hotel.room_type NOT NULL,
     price DECIMAL(10, 2) NOT NULL CHECK (price > 0),
     max_guests INT NOT NULL CHECK (max_guests > 0),
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS hotel.room_types (
 );
 
 CREATE TABLE IF NOT EXISTS hotel.rooms (
-    room_id SERIAL PRIMARY KEY,
+    room_id BIGSERIAL PRIMARY KEY,
     number VARCHAR(10) NOT NULL UNIQUE,
     name VARCHAR(50) NOT NULL,
     description TEXT,
@@ -20,21 +20,21 @@ CREATE TABLE IF NOT EXISTS hotel.rooms (
 );
 
 CREATE TABLE IF NOT EXISTS hotel.amenities (
-    amenity_id SERIAL PRIMARY KEY,
+    amenity_id BIGSERIAL PRIMARY KEY,
     name VARCHAR(100) NOT NULL,
     description TEXT
 );
 
 CREATE TABLE IF NOT EXISTS hotel.room_amenities (
-    room_id INT NOT NULL,
-    amenity_id INT NOT NULL,
+    room_id BIGINT  NOT NULL,
+    amenity_id BIGINT  NOT NULL,
     PRIMARY KEY (room_id, amenity_id),
     FOREIGN KEY (room_id) REFERENCES hotel.rooms(room_id) ON DELETE CASCADE,
     FOREIGN KEY (amenity_id) REFERENCES hotel.amenities(amenity_id) ON DELETE CASCADE
 );
 
 CREATE TABLE IF NOT EXISTS hotel.guests (
-    guest_id SERIAL PRIMARY KEY,
+    guest_id BIGSERIAL PRIMARY KEY,
     first_name VARCHAR(100) NOT NULL,
     last_name VARCHAR(100) NOT NULL,
     middle_name VARCHAR(100),
@@ -46,14 +46,14 @@ CREATE TABLE IF NOT EXISTS hotel.guests (
 );
 
 CREATE TABLE IF NOT EXISTS hotel.bookings (
-    booking_id SERIAL PRIMARY KEY,
-    guest_id INT NOT NULL REFERENCES hotel.guests(guest_id),
-    room_id INT NOT NULL REFERENCES hotel.rooms(room_id),
+    booking_id BIGSERIAL PRIMARY KEY,
+    guest_id BIGINT NOT NULL REFERENCES hotel.guests(guest_id),
+    room_id BIGINT NOT NULL REFERENCES hotel.rooms(room_id),
     check_in_date DATE NOT NULL,
     check_out_date DATE NOT NULL,
     adults_count INT NOT NULL CHECK (adults_count > 0),
     children_count INT DEFAULT 0 CHECK (children_count >= 0),
-    total_price DECIMAL(12, 2) NOT NULL CHECK (total_price >= 0),
+    total_price NUMERIC(38,2) NOT NULL CHECK (total_price >= 0),
     special_requests TEXT,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -64,7 +64,7 @@ CREATE TABLE IF NOT EXISTS hotel.bookings (
 CREATE TYPE hotel.employee_role AS ENUM ('Администратор', 'Горничная', 'Менеджер', 'Портье');
 
 CREATE TABLE IF NOT EXISTS hotel.employees (
-    employee_id SERIAL PRIMARY KEY,
+    employee_id BIGSERIAL PRIMARY KEY,
     first_name VARCHAR(100) NOT NULL,
     last_name VARCHAR(100) NOT NULL,
     middle_name VARCHAR(100),
@@ -79,9 +79,9 @@ CREATE TABLE IF NOT EXISTS hotel.employees (
 );
 
 CREATE TABLE IF NOT EXISTS hotel.room_cleaning (
-    cleaning_id SERIAL PRIMARY KEY,
-    room_id INT NOT NULL,
-    employee_id INT NOT NULL,
+    cleaning_id BIGSERIAL PRIMARY KEY,
+    room_id BIGINT NOT NULL,
+    employee_id BIGINT NOT NULL,
     cleaning_time TIMESTAMP NOT NULL,
     scheduled_time TIME,
     notes TEXT,
